+++ src/steps/enablement.test.ts
@@ -3,25 +3,6 @@ import { ServiceUsageName } from '../google-cloud/types';
 import * as enablement from './enablement';
 import * as serviceUsage from './service-usage/client';
 
-describe('#getUniqueIntegrationConfigProjectsForStepEnablement', () => {
-  test('should return only "main" integration project ID if no "target" project ID is supplied', () => {
-    const config = getMockIntegrationConfig();
-    expect(
-      enablement.getUniqueIntegrationConfigProjectsForStepEnablement(config),
-    ).toEqual(['j1-gc-integration-dev-v2']);
-  });
-
-  test('should return both "main" and "target" integration project IDs if "target" project ID is supplied', () => {
-    const config = getMockIntegrationConfig({
-      projectId: 'my-target-project-id',
-    });
-
-    expect(
-      enablement.getUniqueIntegrationConfigProjectsForStepEnablement(config),
-    ).toEqual(['j1-gc-integration-dev-v2', 'my-target-project-id']);
-  });
-});
-
 describe('#createStepStartState', () => {
   test('should not mark step start state as disabled when the service is enabled', () => {
     expect(
@@ -107,25 +88,29 @@ describe('#getEnabledServiceNames', () => {
     jest.resetAllMocks();
   });
 
-  test('should return "main" project enabled service names if "target" project not specified in config', async () => {
+  test('intersectedEnabledServices should be "main" project enabled service names if "target" project not specified in config', async () => {
     const mockConfig = getMockIntegrationConfig();
     const mockEnabledServiceNames: string[] = [
       'pubsub.googleapis.com',
       'appengine.googleapis.com',
       'dns.googleapis.com',
     ];
+    const mockEnabledServiceData: enablement.EnabledServiceData = {
+      mainProjectEnabledServices: mockEnabledServiceNames,
+      intersectedEnabledServices: mockEnabledServiceNames,
+    };
 
     const collectEnabledServicesForProjectSpy = jest
       .spyOn(serviceUsage, 'collectEnabledServicesForProject')
       .mockResolvedValue(Promise.resolve(mockEnabledServiceNames));
 
     expect(await enablement.getEnabledServiceNames(mockConfig)).toEqual(
-      mockEnabledServiceNames,
+      mockEnabledServiceData,
     );
     expect(collectEnabledServicesForProjectSpy).toHaveBeenCalledTimes(1);
   });
 
-  test('should return "main" project enabled service names if "target" project specified in config is equal to "main" project', async () => {
+  test('intersectedEnabledServices should be "main" project enabled service names if "target" project specified in config is equal to "main" project', async () => {
     const mockConfig = getMockIntegrationConfig({
       projectId: 'j1-gc-integration-dev-v2',
     });
@@ -135,18 +120,26 @@ describe('#getEnabledServiceNames', () => {
       'appengine.googleapis.com',
       'dns.googleapis.com',
     ];
+    const targetMockEnabledServiceNames: string[] = [
+      'appengine.googleapis.com',
+    ];
+    const mockEnabledServiceData: enablement.EnabledServiceData = {
+      mainProjectEnabledServices: mockEnabledServiceNames,
+      intersectedEnabledServices: mockEnabledServiceNames,
+    };
 
     const collectEnabledServicesForProjectSpy = jest
       .spyOn(serviceUsage, 'collectEnabledServicesForProject')
-      .mockResolvedValue(Promise.resolve(mockEnabledServiceNames));
+      .mockResolvedValueOnce(Promise.resolve(mockEnabledServiceNames))
+      .mockResolvedValueOnce(Promise.resolve(targetMockEnabledServiceNames));
 
     expect(await enablement.getEnabledServiceNames(mockConfig)).toEqual(
-      mockEnabledServiceNames,
+      mockEnabledServiceData,
     );
     expect(collectEnabledServicesForProjectSpy).toHaveBeenCalledTimes(1);
   });
 
-  test('should return enabled service names intersection between "main" and "target" projects if "target" project specified in config', async () => {
+  test('intersectedEnabledServices should be the enabled service names intersection between "main" and "target" projects if "target" project specified in config', async () => {
     const mockConfig = getMockIntegrationConfig({
       projectId: 'my-target-project-id',
     });
@@ -160,15 +153,20 @@ describe('#getEnabledServiceNames', () => {
     const targetMockEnabledServiceNames: string[] = [
       'appengine.googleapis.com',
     ];
+    const mockEnabledServiceData: enablement.EnabledServiceData = {
+      mainProjectEnabledServices: mainMockEnabledServiceNames,
+      targetProjectEnabledServices: targetMockEnabledServiceNames,
+      intersectedEnabledServices: targetMockEnabledServiceNames,
+    };
 
     const collectEnabledServicesForProjectSpy = jest
       .spyOn(serviceUsage, 'collectEnabledServicesForProject')
-      .mockResolvedValue(Promise.resolve(mainMockEnabledServiceNames))
-      .mockResolvedValue(Promise.resolve(targetMockEnabledServiceNames));
+      .mockResolvedValueOnce(Promise.resolve(mainMockEnabledServiceNames))
+      .mockResolvedValueOnce(Promise.resolve(targetMockEnabledServiceNames));
 
-    expect(await enablement.getEnabledServiceNames(mockConfig)).toEqual([
-      'appengine.googleapis.com',
-    ]);
+    expect(await enablement.getEnabledServiceNames(mockConfig)).toEqual(
+      mockEnabledServiceData,
+    );
 
     expect(collectEnabledServicesForProjectSpy).toHaveBeenCalledTimes(2);
   });